fix: spec
