+++ docx-core/src/reader/table_cell.rs
@@ -5,7 +5,6 @@ use xml::attribute::OwnedAttribute;
 use xml::reader::{EventReader, XmlEvent};
 
 use super::*;
-use crate::types::*;
 
 impl ElementReader for TableCell {
     fn read<R: Read>(r: &mut EventReader<R>, _: &[OwnedAttribute]) -> Result<Self, ReaderError> {
@@ -25,86 +24,12 @@ impl ElementReader for TableCell {
                             cell = cell.add_paragraph(p);
                             continue;
                         }
-                        XMLElement::TableCellProperty => loop {
-                            let e = r.next();
-                            match e {
-                                Ok(XmlEvent::StartElement {
-                                    attributes, name, ..
-                                }) => {
-                                    let e = XMLElement::from_str(&name.local_name).unwrap();
-
-                                    ignore::ignore_element(
-                                        e.clone(),
-                                        XMLElement::TableCellPropertyChange,
-                                        r,
-                                    );
-
-                                    match e {
-                                        XMLElement::TableCellWidth => {
-                                            let mut w = 0;
-                                            let mut width_type = WidthType::Auto;
-                                            for a in attributes {
-                                                let local_name = &a.name.local_name;
-                                                if local_name == "type" {
-                                                    width_type = WidthType::from_str(&a.value)?;
-                                                } else if local_name == "w" {
-                                                    w = usize::from_str(&a.value)?;
-                                                }
-                                            }
-                                            cell = cell.width(w, width_type);
-                                        }
-                                        XMLElement::TableGridSpan => {
-                                            if let Some(a) = &attributes.get(0) {
-                                                cell = cell.grid_span(usize::from_str(&a.value)?)
-                                            }
-                                        }
-                                        XMLElement::TableVMerge => {
-                                            if let Some(a) = &attributes.get(0) {
-                                                cell = cell.vertical_merge(VMergeType::from_str(
-                                                    &a.value,
-                                                )?);
-                                            } else {
-                                                // Treat as a continue without attribute
-                                                cell = cell.vertical_merge(VMergeType::Continue)
-                                            }
-                                        }
-                                        XMLElement::VAlign => {
-                                            if let Some(a) = &attributes.get(0) {
-                                                cell = cell.vertical_align(VAlignType::from_str(
-                                                    &a.value,
-                                                )?);
-                                            }
-                                        }
-                                        XMLElement::Shading => {
-                                            if let Ok(shd) = Shading::read(r, &attributes) {
-                                                cell = cell.shading(shd);
-                                            }
-                                        }
-                                        XMLElement::TextDirection => {
-                                            if let Some(a) = &attributes.get(0) {
-                                                if let Ok(v) = TextDirectionType::from_str(&a.value)
-                                                {
-                                                    cell = cell.text_direction(v);
-                                                }
-                                            }
-                                        }
-                                        XMLElement::TableCellBorders => {
-                                            let borders = TableCellBorders::read(r, &attributes)?;
-                                            cell = cell.set_borders(borders);
-                                        }
-                                        _ => {}
-                                    }
-                                }
-                                Ok(XmlEvent::EndElement { name, .. }) => {
-                                    let e = XMLElement::from_str(&name.local_name).unwrap();
-                                    if e == XMLElement::TableCellProperty {
-                                        break;
-                                    }
-                                }
-                                Err(_) => return Err(ReaderError::XMLReadError),
-                                _ => {}
+                        XMLElement::TableCellProperty => {
+                            if let Ok(p) = TableCellProperty::read(r, &attributes) {
+                                cell.property = p;
                             }
-                        },
+                            continue;
+                        }
                         _ => {}
                     }
                 }
@@ -125,6 +50,7 @@ impl ElementReader for TableCell {
 mod tests {
 
     use super::*;
+    use crate::types::*;
     #[cfg(test)]
     use pretty_assertions::assert_eq;
