+++ docx-wasm/js/json/styles.ts
@@ -1,6 +1,7 @@
 import { RunPropertyJSON } from "./run";
 import { ParagraphPropertyJSON } from "./paragraph";
 import { TablePropertyJSON } from "./table";
+import { TableCellPropertyJSON } from "..";
 
 export type StyleJSON = {
   styleId: string;
